+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Increases unit-tests coverage of core components
 - Improves threads management
 - Improves TableWidget performance
+- Fixes #7 - Checks for running experiments before closing a project
 
 * Thu Jun 21 2018 Marcos Cardinot <marcos@cardinot.net> - 0.1.0-alpha0
 - First public release